refactor: move otlp trace service into otlp feature
